chore: set up basic routes
