+++ packages/frontend/src/pages/posts/[postId].tsx
