+++ packages/frontend/src/pages/users.tsx
@@ -1,0 +1,3 @@
+export default function Users() {
+  return <div>Users</div>;
+}
