+++ packages/frontend/src/pages/users/[userId].tsx
@@ -1,0 +1,8 @@
+import { useRouter } from 'next/router';
+
+export default function User() {
+  const router = useRouter();
+  const { userId } = router.query;
+
+  return <div>User {userId}</div>;
+}
